+++ frontend/src/App.tsx
@@ -17,6 +17,7 @@ import './App.css'
 function App() {
   return (
     <div className="App">
+      
       <AuthContextProvider>
         <HashRouter> 
         <Routes>
