+++ frontend/src/pages/Login.css
@@ -5,7 +5,8 @@ form {
   justify-content: center;
   margin: auto;
   margin-top: 100px;
-  padding-bottom: 20px;
+  padding-top: 40px;
+  padding-bottom: 170px;
 }
 
 label {
