+++ frontend/src/pages/Results.css
@@ -6,6 +6,14 @@
   align-items: center;
 }
 
+.warning {
+  font-size: 40px;
+  align-items: center;
+  text-align: center;
+  margin: 0, auto;
+  padding-bottom: 600px;
+}
+
 #heading {
   font-size: 60px;
   font-weight: 700;
